perf: 优化

+++ src/components/CustomForm/VUpload/Upload.vue
@@ -2,46 +2,47 @@
   <el-upload
     ref="upload"
     class="oms-upload"
+    multiple
+    list-type="picture-card"
     :headers="headers"
+    :limit="limit"
     :class="{hide:showUpload}"
     :action="action"
-    list-type="picture-card"
     :auto-upload="true"
-    multiple
     :before-upload="beforeUpload"
+    :before-remove="beforeRemove"
     :on-success="handleSuccess"
     :on-change="handleChange"
-    :limit="limit"
     :on-exceed="onExceed"
   >
     <!-- 上传 -->
-    <i slot="default" class="el-icon-upload2 upload-icon" />
+    <i slot="default" class="el-icon-upload2 upload-icon" :style="{ width: `${width - 2}px`, height: `${height - 2}px`, lineHeight: `${height - 2}px`}" />
     <div slot="tip"><slot name="tip" /></div>
     <!-- 自定义文件展示区 -->
-    <div slot="file" slot-scope="{file}" class="file-slot">
+    <div slot="file" slot-scope="{file}" class="file-slot" :style="{ width: `${width - 2}px`, height: `${height - 2}px`}">
       <el-image
         v-if="file.url"
         :ref="`image_${file.uid}`"
         class="el-upload-list__item-thumbnail"
         lazy
         :src="file.url"
-        :preview-src-list="[file.url]"
+        :preview-src-list="getImagePreviewList(value, file.url)"
       >
-        <div slot="error" class="file-icon">
+        <div slot="error" class="file-icon" :style="{lineHeight: `${height - 2}px`}">
           <i class="el-icon-document" />
         </div>
       </el-image>
 
       <el-progress
         v-if="file.percentage !== undefined && file.status !== 'success'"
-        :width="66"
-        :height="66"
+        :width="width > height ? height - 2 : width - 2"
+        :style="{ width: `${width}px`, height: `${height}px`}"
         class="progress"
         :type="'circle'"
         :stroke-width="6"
         :percentage="parsePercentage(file.percentage)"
       />
-      <!-- 操作按钮组 -->
+      <!-- 操作按钮组-->
       <span v-else class="el-upload-list__item-actions">
         <div class="actions">
           <i class="el-icon-zoom-in preview" @click="handlePreview(file)" />
@@ -63,15 +64,23 @@ function toArray(value) {
 export default {
   props: {
     value: { type: [Array, String], default: '' },
+    width: { type: Number, default: 70 },
+    height: { type: Number, default: 70 },
     headers: { type: Object, default: _ => {} },
     action: { type: String, default: process.env.VUE_APP_BASE_API + '/upload/image' },
     // 是否可下载
-    downloadable: { type: Boolean, default: _ => false },
+    downloadable: { type: Boolean, default: false },
+    // 删除前二次确认
+    removeConfirm: { type: Boolean, default: false },
+    // 删除前二次确认提示文本
+    removeConfirmText: { type: String, default: '此操作将删除该文件, 是否继续?' },
     // 上传数量
     limit: { type: Number, default: 1 },
     accept: { type: String, default: '' },
     format: { type: Array, default: _ => [] },
-    maxSize: { type: Number, default: 0 }
+    maxSize: { type: Number, default: 0 },
+    // 删除前置钩子
+    beforeRemove: { type: Function, default: () => {} }
   },
   data() {
     return {
@@ -139,10 +148,25 @@ export default {
         return false
       }
     },
-    // remove操作
+    // 点击 remove 操作
     removeFile(file) {
-      this.$refs.upload.handleRemove(file)
-      this.updateValue()
+      // 删除二次确认
+      if (this.removeConfirm) {
+        this.$confirm(this.removeConfirmText, '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning',
+          center: false
+        }).then(() => {
+          // handleRemove 里会调用 beforeRemove 钩子
+          this.$refs.upload.handleRemove(file)
+          this.updateValue()
+        }).catch(() => {
+        })
+      } else {
+        this.$refs.upload.handleRemove(file)
+        this.updateValue()
+      }
     },
     // 预览操作
     handlePreview(file) {
@@ -172,6 +196,7 @@ export default {
         return
       }
       const url = response.payload.file_path
+      this.$emit('success', response.payload)
       this.$notify.success({
         title: '上传成功',
         message: file.name
@@ -189,6 +214,22 @@ export default {
     },
     parsePercentage(val = 0) {
       return parseInt(val, 10)
+    },
+    // 将当前图片的 index 设为0
+    getImagePreviewList(value, currentImg) {
+      const imgList = toArray(value)
+      let pervFlag = true; const pervArr = []; const nextArr = []
+      imgList.map(img => {
+        if (img !== currentImg && pervFlag) {
+          pervArr.push(img)
+        } else {
+          if (img === currentImg) {
+            pervFlag = false
+          }
+          nextArr.push(img)
+        }
+      })
+      return [...nextArr, ...pervArr]
     }
   }
 }
@@ -224,13 +265,10 @@ export default {
   }
   .progress{
     background: #fff;
-    width: 70px;
-    height: 70px;
+
     padding: 2px;
   }
   .file-slot{
-    width: 68px;
-    height: 68px;
     line-height: initial;
     vertical-align: middle;
   }
@@ -238,9 +276,9 @@ export default {
 /deep/ {
   .el-upload-list--picture-card .el-upload-list__item,
   .el-upload--picture-card{
-    width: 70px;
-    height: 70px;
-    line-height: 68px;
+    width: auto;
+    height: auto;
+    line-height: initial;
     margin-bottom: 8px;
   }
   .el-upload-list--picture-card{
@@ -257,12 +295,16 @@ export default {
     // 关掉动画 避免闪动
     transition: none;
     &.el-list-enter,&.el-list-leave-active{-webkit-transform:translateY(0px);transform:translateY(0px)}
+
+    // 去除取消删除之后的外边框线
+    &.focusing {
+      outline-width:0
+    }
   }
 
   .file-icon{
     font-size: 20px;
     text-align: center;
-    line-height: 68px;
   }
   // 垂直居中 等比缩放
   .el-image__inner{
@@ -278,5 +320,12 @@ export default {
 }
 .hide/deep/ .el-upload {
   display: none;
+}
+// 使进度条环 居中
+/deep/ .el-progress-circle{
+  position: absolute;
+  left: 50%;
+  top: 50%;
+  transform: translate(-50%, -50%);
 }
 </style>
